Moved GameUpdateMessage type to gameutility
